BC-1777 Refactor use case
